% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{Frequency table of large dataframe}
\usage{
freq_table(data, ordered = TRUE)
}
\arguments{
\item{data}{A matrix or a dataframe where each column is a variable.}

\item{ordered}{A logical indicating whether to order the output matrix
based on the decreasing ordering of the cell frequencies.}
}
\value{
A matrix with all the possible combinations of the variables
and a column with the frequency of each combination.
}
\description{
Compute a summarized matrix of the contingency table of the dataframe.
}
\examples{
set.seed(1)
x <- sample(c(0,1), 100, replace = T)
y <- sample(c(0,1), 100, replace = T)
z <- sample(c(0,1), 100, replace = T)
df <- cbind(x, y, z)
freq_table(df)

}
\keyword{internal}
