% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFormula.R
\name{mFormula}
\alias{mFormula}
\title{Formula}
\usage{
mFormula(adjm)
}
\arguments{
\item{adjm}{an adjacency matrix.}
}
\value{
A formula for the model.
}
\description{
Define the model formula from an adjacency matrix.
}
\keyword{internal}
