% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MorbiGraph-package.R
\docType{package}
\name{MorbiGraph-package}
\alias{MorbiGraph}
\alias{MorbiGraph-package}
\title{MorbiGraph: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
