% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_stepwise.R
\name{node_stepwise}
\alias{node_stepwise}
\title{Node-wise stepwise selection}
\usage{
node_stepwise(
  data,
  v.conf = NULL,
  direction = "backward",
  IC = "EBIC",
  rule = "AND"
)
}
\arguments{
\item{data}{a matrix or a dataframe containing all the variables. All columns must be named.}

\item{v.conf}{A string with the names of the confounding variables. By design, the associations between them will be zero.}

\item{direction}{the mode of stepwise search, can be one of "backward" or "forward".}

\item{IC}{information criteria to use: "AIC", "BIC", "EBIC".}

\item{rule}{Can be "AND" or "OR" to indicate whether the AND-rule or the OR-rule should be used to define the edges in the network.}
}
\value{
The function returns the estimated adjacency matrix of the graph.
}
\description{
Structure learning of the graph achieved through stepwise selection.
}
\examples{

set.seed(1)
df <- matrix(sample(0:1,100,replace=T),ncol=5)
colnames(df) <- paste0('X',1:5)

net <- node_stepwise(df)

}
