% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR.R
\name{OR}
\alias{OR}
\title{Odds ratio of dichotomous variables}
\usage{
OR(data, test = FALSE)
}
\arguments{
\item{data}{A matrix where each column is a variable.}

\item{test}{a logical indicating whether to perform chi-squared test for contingency tables.}
}
\value{
A symmetric matrix of odds ratios.
For each pair of variables, it creates a 2x2 contingency table and then it computes the odds ratio.
If \code{test=TRUE} it returns the pvalues for the test computed for each 2x2 table.
}
\description{
Odds ratio of dichotomous variables
}
\examples{

set.seed(1)
x <- sample(c(0,1), 100, replace = T)
y <- sample(c(0,1), 100, replace = T)
z <- sample(c(0,1), 100, replace = T)
df <- cbind(x, y, z)

or <- OR(df)

}
