% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_tab.R
\name{OR_tab}
\alias{OR_tab}
\title{Odds ratio for two dichotomous variables}
\usage{
OR_tab(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A number. First it creates a 2x2 table and then it computes the odds ratio.
}
\description{
Odds ratio for two dichotomous variables
}
\examples{
set.seed(1)
x <- sample(c(0,1), 100, replace = T)
y <- sample(c(0,1), 100, replace = T)
OR_tab(x,y)

}
\keyword{internal}
