% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MorbiGraph-package.R
\docType{package}
\name{MorbiGraph-package}
\alias{MorbiGraph}
\alias{MorbiGraph-package}
\title{MorbiGraph: Graphical models for multimorbidity pattern analysis}
\description{
Graphical models for multimorbidity pattern analysis
}
\keyword{internal}
