% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qYule.R
\name{qYule}
\alias{qYule}
\title{Yule's Q}
\usage{
qYule(x, OR = TRUE)
}
\arguments{
\item{x}{A symmetric matrix of odds ratios or a dataframe with variables
for which to compute the coefficients.}

\item{OR}{logical indicating if the input is a matrix of odds ratios (TRUE) or a dataframe (FALSE).}
}
\value{
A symmetric matrix with the Q coefficients computed from the OR or the dataframe.

Values of Q can vary between -1 and 1.
−1 reflects total negative association,
+1 reflects perfect positive association,
and 0 reflects no association.
}
\description{
Compute the Yule coefficient of association.
}
\details{
The coefficient is computed as
Q = (OR-1)/(OR+1),
where OR is the odds ratio.
}
\examples{

# single value
qYule(1.5)

# matrix of odds ratios
OR <- matrix(c(1, 3.5, 0.1,
               3.5, 1, 1.3,
               0.1, 1.3, 1), ncol=3)
qYule(OR)

}
